src/edge_node/mqtt_publisher.py
"""
MQTT publisher for edge-to-cloud communication.

Implements reliable MQTT publishing with QoS guarantees for real-time
biosignal and pose estimation data transmission to cloud fusion backend.

Author: Dr. Li Chen
Date: 2024-01-10
"""

import json
import logging
from typing import Dict, Any, Optional
from dataclasses import asdict
import paho.mqtt.client as mqtt


logger = logging.getLogger(__name__)


class MQTTPublisher:
    """MQTT publisher for streaming edge inference results to cloud.
    
    Provides reliable message publishing with automatic reconnection,
    QoS level 1 (at-least-once delivery), and message buffering during
    connectivity interruptions.
    
    Topics:
        - bio-resilience/edge/{device_id}/pose: Pose estimation results
        - bio-resilience/edge/{device_id}/activity: Activity classifications
        - bio-resilience/edge/{device_id}/metrics: Performance metrics
        - bio-resilience/edge/{device_id}/status: Device health status
    
    Args:
        broker: MQTT broker hostname or IP address
        port: MQTT broker port (default: 1883)
        device_id: Unique edge device identifier
        qos: Quality of Service level (0, 1, or 2)
        keepalive: Connection keepalive interval in seconds
        
    Example:
        >>> publisher = MQTTPublisher(broker="mqtt.bio-resilience.org", device_id="jetson_001")
        >>> publisher.connect()
        >>> publisher.publish_pose_data({"keypoints": [...], "timestamp": 1704067200})
    """
    
    def __init__(
        self,
        broker: str,
        port: int = 1883,
        device_id: str = "edge_node_default",
        qos: int = 1,
        keepalive: int = 60
    ):
        """Initialize MQTT publisher."""
        self.broker = broker
        self.port = port
        self.device_id = device_id
        self.qos = qos
        self.keepalive = keepalive
        
        # Initialize MQTT client
        self.client = mqtt.Client(client_id=f"edge_{device_id}")
        self.client.on_connect = self._on_connect
        self.client.on_disconnect = self._on_disconnect
        self.client.on_publish = self._on_publish
        
        # Connection state
        self.connected = False
        
        # Message buffer for offline periods
        self.message_buffer = []
        self.max_buffer_size = 1000
        
        logger.info(f"Initialized MQTT publisher for device {device_id}")
    
    def connect(self) -> bool:
        """Establish connection to MQTT broker.
        
        Returns:
            True if connection successful, False otherwise
        """
        pass
    
    def disconnect(self) -> None:
        """Gracefully disconnect from MQTT broker."""
        pass
    
    def publish_pose_data(
        self,
        pose_data: Dict[str, Any],
        timestamp: float
    ) -> None:
        """Publish pose estimation results to cloud.
        
        Args:
            pose_data: Dictionary containing keypoints, bounding boxes, etc.
            timestamp: Data timestamp in Unix epoch seconds
        """
        pass
    
    def publish_activity_data(
        self,
        activity_data: Dict[str, Any],
        timestamp: float
    ) -> None:
        """Publish activity classification results.
        
        Args:
            activity_data: Activity labels and confidence scores
            timestamp: Data timestamp
        """
        pass
    
    def publish_metrics(self, metrics: Dict[str, float]) -> None:
        """Publish performance metrics to monitoring topic.
        
        Args:
            metrics: Dictionary of performance metrics (FPS, latency, etc.)
        """
        pass
    
    def _on_connect(
        self,
        client: mqtt.Client,
        userdata: Any,
        flags: Dict,
        rc: int
    ) -> None:
        """Callback for successful MQTT connection."""
        pass
    
    def _on_disconnect(
        self,
        client: mqtt.Client,
        userdata: Any,
        rc: int
    ) -> None:
        """Callback for MQTT disconnection."""
        pass
    
    def _on_publish(
        self,
        client: mqtt.Client,
        userdata: Any,
        mid: int
    ) -> None:
        """Callback for successful message publish."""
        pass
    
    def _flush_buffer(self) -> None:
        """Flush buffered messages after reconnection."""
        pass
