src/edge_node/video_capture.py
"""
Hardware-accelerated video capture for edge devices.

Optimized video capture implementation leveraging GStreamer pipeline
on NVIDIA Jetson for efficient camera interfacing and RTSP streaming.

Author: Dr. Li Chen
Date: 2024-01-14
"""

import logging
from typing import Optional, Tuple
import cv2
import numpy as np


logger = logging.getLogger(__name__)


class VideoCapture:
    """Hardware-accelerated video capture wrapper.
    
    Provides unified interface for camera capture, RTSP streams, and video files
    with automatic GStreamer pipeline optimization on Jetson platforms.
    
    Args:
        source: Video source (camera index, RTSP URL, or file path)
        width: Desired frame width (None for native resolution)
        height: Desired frame height (None for native resolution)
        fps: Target frames per second
        use_gstreamer: Enable GStreamer hardware acceleration
        
    Example:
        >>> capture = VideoCapture(source=0, width=1920, height=1080, fps=30)
        >>> frame = capture.read()
    """
    
    def __init__(
        self,
        source: int | str = 0,
        width: Optional[int] = None,
        height: Optional[int] = None,
        fps: int = 30,
        use_gstreamer: bool = True
    ):
        """Initialize video capture."""
        self.source = source
        self.width = width
        self.height = height
        self.fps = fps
        self.use_gstreamer = use_gstreamer
        
        # Initialize capture
        self.cap = self._initialize_capture()
        
        logger.info(f"Initialized video capture from source: {source}")
    
    def _initialize_capture(self) -> cv2.VideoCapture:
        """Initialize OpenCV VideoCapture with optimal settings.
        
        Returns:
            Configured cv2.VideoCapture object
        """
        pass
    
    def _build_gstreamer_pipeline(self) -> str:
        """Build GStreamer pipeline for hardware-accelerated capture.
        
        Returns:
            GStreamer pipeline string
        """
        pass
    
    def read(self) -> Optional[np.ndarray]:
        """Read next frame from video source.
        
        Returns:
            Frame as BGR numpy array, or None if capture failed
        """
        pass
    
    def get_properties(self) -> dict:
        """Get current capture properties.
        
        Returns:
            Dictionary of capture properties (width, height, fps, etc.)
        """
        pass
    
    def release(self) -> None:
        """Release video capture resources."""
        pass
    
    def __enter__(self):
        """Context manager entry."""
        return self
    
    def __exit__(self, exc_type, exc_val, exc_tb):
        """Context manager exit."""
        self.release()
