requirements.txt
# Core Deep Learning
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
ultralytics==8.1.24
onnx==1.15.0
onnxruntime-gpu==1.16.3

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
mediapipe==0.10.9
albumentations==1.3.1
scikit-image==0.22.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2

# Backend & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
httpx==0.26.0

# Message Queue & Streaming
paho-mqtt==1.6.1
kafka-python==2.0.2
aiokafka==0.10.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
sentry-sdk==1.39.2

# Signal Processing
kalman-filter==0.5.0
pykalman==0.9.7
filterpy==1.4.5
biosignalsnotebooks==1.2.1

# Time Series Analysis
statsmodels==0.14.1
pmdarima==2.0.4
prophet==1.1.5

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.3

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
pillow==10.2.0
requests==2.31.0
aiofiles==23.2.1

# Edge-specific (NVIDIA Jetson)
# jetson-stats==4.2.6  # Uncomment for Jetson deployment
# tensorrt==8.6.1.6    # Requires NVIDIA JetPack SDK

# WatchOS Integration (Python bridge)
bleak==0.21.1
pygatt==4.0.5

# Database
databases==0.8.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Security
cryptography==42.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
